Refactor postprocessing folder
