+++ mlreco/post_processing/__init__.py
@@ -1,16 +1,18 @@
-from .store_uresnet import store_uresnet
-from .store_uresnet_ppn import store_uresnet_ppn
-from .store_input import store_input
-from .deghosting_metrics import deghosting_metrics
-from .michel_reconstruction import michel_reconstruction
-from .track_clustering import track_clustering
-from .instance_clustering import instance_clustering
-from .michel_reconstruction_2d import michel_reconstruction_2d
-from .uresnet_metrics import uresnet_metrics
-from .michel_reconstruction_noghost import michel_reconstruction_noghost
-from .ppn_metrics import ppn_metrics
-from .cluster_gnn_metrics import cluster_gnn_metrics
-from .cluster_cnn_metrics import cluster_cnn_metrics
-from .cosmic_discriminator_metrics import cosmic_discriminator_metrics
-from .kinematics_metrics import kinematics_metrics
-from .vertex_metrics import vertex_metrics
+from .analysis.instance_clustering import instance_clustering
+from .analysis.michel_reconstruction_2d import michel_reconstruction_2d
+from .analysis.michel_reconstruction_noghost import michel_reconstruction_noghost
+from .analysis.michel_reconstruction import michel_reconstruction
+from .analysis.track_clustering import track_clustering
+
+from .metrics.cluster_cnn_metrics import cluster_cnn_metrics
+from .metrics.cluster_gnn_metrics import cluster_gnn_metrics
+from .metrics.cosmic_discriminator_metrics import cosmic_discriminator_metrics
+from .metrics.deghosting_metrics import deghosting_metrics
+from .metrics.kinematics_metrics import kinematics_metrics
+from .metrics.ppn_metrics import ppn_metrics
+from .metrics.uresnet_metrics import uresnet_metrics
+from .metrics.vertex_metrics import vertex_metrics
+
+from .store.store_input import store_input
+from .store.store_uresnet_ppn import store_uresnet_ppn
+from .store.store_uresnet import store_uresnet
