+++ mlreco/post_processing/analysis/michel_reconstruction_noghost.py
